+++ ms2query/run_ms2query.py
@@ -7,9 +7,11 @@ from ms2query.utils import convert_files_to_matchms_spectrum_objects
 from urllib.request import urlretrieve
 
 
-def default_library_file_names() -> Dict[str, str]:
-    """Returns a dictionary with the file names of default files for a MS2Library"""
+def default_library_file_base_names() -> Dict[str, str]:
+    """Returns a dictionary with the base names of default files for a MS2Library"""
     return {"sqlite": "ALL_GNPS_210409_train_split.sqlite",
+            "sqlite_trainables": "ALL_GNPS_210409_Spec2Vec_ms2query.model.trainables.syn1neg.npy",
+            "sqlite_vectors": "ALL_GNPS_210409_Spec2Vec_ms2query.model.wv.vectors.npy",
             "classifiers": "ALL_GNPS_210409_positive_processed_annotated_CF_NPC_classes.txt",
             "s2v_model": "ALL_GNPS_210409_Spec2Vec_ms2query.model",
             "ms2ds_model": "ms2ds_20210420-141937_data210409_10k_500_500_200.hdf5",
@@ -18,39 +20,8 @@ def default_library_file_names() -> Dict[str, str]:
             "ms2ds_embeddings": "ms2ds_embeddings_train_spectra_210426.pickle"}
 
 
-def create_default_library_object(directory: str,
-                                  file_name_dictionary: Dict[str, str]
-                                  ) -> MS2Library:
-    """Creates a library object for specified directory and file names
-
-    For default file names the function default_library_file_names can be run.
-
-    Args:
-    ------
-    directory:
-        Path to the directory in which the files are stored
-    file_name_dictionary:
-        A dictionary with as keys the type of file and as values the file names
-    """
-    sqlite_file = os.path.join(directory, file_name_dictionary["sqlite"])
-    classifiers_file = os.path.join(directory, file_name_dictionary["classifiers"])
-
-    # Models
-    s2v_model_file = os.path.join(directory, file_name_dictionary["s2v_model"])
-    ms2ds_model_file = os.path.join(directory, file_name_dictionary["ms2ds_model"])
-    ms2query_model = os.path.join(directory, file_name_dictionary["ms2query_model"])
-
-    # Embeddings
-    s2v_embeddings_file = os.path.join(directory, file_name_dictionary["s2v_embeddings"])
-    ms2ds_embeddings_file = os.path.join(directory, file_name_dictionary["ms2ds_embeddings"])
-
-    return MS2Library(sqlite_file, s2v_model_file, ms2ds_model_file,
-                      s2v_embeddings_file, ms2ds_embeddings_file,
-                      ms2query_model, classifiers_file)
-
-
-def automatically_download_models(dir_to_store_files: str,
-                                  file_name_dict: Dict[str, str]):
+def download_default_models(dir_to_store_files: str,
+                            file_name_dict: Dict[str, str]):
     """Downloads files from Zenodo
 
     Args:
@@ -63,11 +34,11 @@ def automatically_download_models(dir_to_store_files: str,
     """
     if not os.path.exists(dir_to_store_files):
         os.mkdir(dir_to_store_files)
-    zenodo_files_location = "https://zenodo.org/record/5564815/files/"
+    zenodo_files_location = "https://zenodo.org/record/5645246/files/"
     for file_name in tqdm(file_name_dict.values(),
                           "Downloading library files"):
         complete_url = zenodo_files_location + file_name + "?download=1"
-        file_location = dir_to_store_files + file_name
+        file_location = os.path.join(dir_to_store_files, file_name)
         if not os.path.exists(file_location):
             urlretrieve(complete_url, file_location)
         else:
@@ -105,27 +76,33 @@ def run_complete_folder(ms2library: MS2Library,
     library_search:
         If True a library search (Finding true matches) is performed
     """
+    # pylint: disable=too-many-arguments
+
     # check if there is a results folder otherwise create one
     if not os.path.exists(results_folder):
         os.mkdir(results_folder)
+    if analog_search:
+        if not os.path.exists(os.path.join(results_folder, "analog_search")):
+            os.mkdir(os.path.join(results_folder, "analog_search"))
+    if library_search:
+        if not os.path.exists(os.path.join(results_folder, "library_search")):
+            os.mkdir(os.path.join(results_folder, "library_search"))
+
     # Go through spectra files in directory
     for file_name in os.listdir(folder_with_spectra):
         file_path = os.path.join(folder_with_spectra, file_name)
         # skip folders
         if os.path.isfile(file_path):
             spectra = convert_files_to_matchms_spectrum_objects(os.path.join(folder_with_spectra, file_name))
-            add_unknown_charges_to_spectra(spectra)
-            if analog_search:
-                if not os.path.exists(os.path.join(results_folder, "analog_search")):
-                    os.mkdir(os.path.join(results_folder, "analog_search"))
-                analogs_results_file_name = os.path.join(results_folder, "analog_search", os.path.splitext(file_name)[0] + ".csv")
-                ms2library.analog_search_store_in_csv(spectra,
-                                                      analogs_results_file_name,
-                                                      nr_of_top_analogs_to_save=nr_of_analogs_to_store,
-                                                      minimal_ms2query_metascore=minimal_ms2query_score)
-            if library_search:
-                if not os.path.exists(os.path.join(results_folder, "library_search")):
-                    os.mkdir(os.path.join(results_folder, "library_search"))
-                library_results_file_name = os.path.join(results_folder, "library_search", os.path.splitext(file_name)[0] + ".csv")
-                ms2library.store_potential_true_matches(spectra,
-                                                        library_results_file_name)
+            if spectra is not None:
+                add_unknown_charges_to_spectra(spectra)
+                if analog_search:
+                    analogs_results_file_name = os.path.join(results_folder, "analog_search", os.path.splitext(file_name)[0] + ".csv")
+                    ms2library.analog_search_store_in_csv(spectra,
+                                                          analogs_results_file_name,
+                                                          nr_of_top_analogs_to_save=nr_of_analogs_to_store,
+                                                          minimal_ms2query_metascore=minimal_ms2query_score)
+                if library_search:
+                    library_results_file_name = os.path.join(results_folder, "library_search", os.path.splitext(file_name)[0] + ".csv")
+                    ms2library.store_potential_true_matches(spectra,
+                                                            library_results_file_name)