+++ tests/test_ms2library.py
@@ -7,8 +7,8 @@ import pandas as pd
 from pandas.testing import assert_frame_equal
 from matchms import Spectrum
 from tensorflow.keras.models import load_model as load_nn_model
-from ms2query.ms2library import MS2Library
-from ms2query.ms2library import get_ms2query_model_prediction_single_spectrum
+from ms2query.ms2library import MS2Library, get_ms2query_model_prediction_single_spectrum, \
+    create_library_object_from_one_dir
 from ms2query.utils import load_pickled_file
 from tests.test_utils import create_test_classifier_csv_file
 from ms2query.results_table import ResultsTable
@@ -95,14 +95,9 @@ def test_ms2library_set_settings(file_names):
         ms2ds_model_file_name, ms2ds_embeddings_file_name, \
         spectrum_id_column_name, ms2q_model_file_name = file_names
 
-    test_library = MS2Library(sqlite_file_loc,
-                              spec2vec_model_file_loc,
-                              ms2ds_model_file_name,
-                              s2v_pickled_embeddings_file,
-                              ms2ds_embeddings_file_name,
-                              ms2q_model_file_name,
-                              spectrum_id_column_name=spectrum_id_column_name,
-                              cosine_score_tolerance=0.2)
+    test_library = MS2Library(sqlite_file_loc, spec2vec_model_file_loc, ms2ds_model_file_name,
+                              s2v_pickled_embeddings_file, ms2ds_embeddings_file_name, ms2q_model_file_name,
+                              spectrum_id_column_name=spectrum_id_column_name, cosine_score_tolerance=0.2)
 
     assert test_library.settings["cosine_score_tolerance"] == 0.2, \
         "Different value for attribute was expected"
@@ -120,12 +115,8 @@ def test_select_potential_true_matches(file_names, test_spectra):
         ms2ds_model_file_name, ms2ds_embeddings_file_name, \
         spectrum_id_column_name, ms2q_model_file_name = file_names
 
-    test_library = MS2Library(sqlite_file_loc,
-                              spec2vec_model_file_loc,
-                              ms2ds_model_file_name,
-                              s2v_pickled_embeddings_file,
-                              ms2ds_embeddings_file_name,
-                              ms2q_model_file_name,
+    test_library = MS2Library(sqlite_file_loc, spec2vec_model_file_loc, ms2ds_model_file_name,
+                              s2v_pickled_embeddings_file, ms2ds_embeddings_file_name, ms2q_model_file_name,
                               spectrum_id_column_name=spectrum_id_column_name)
 
     results = \
@@ -158,13 +149,9 @@ def test_store_potential_true_matches(file_names, test_spectra, tmp_path):
         spectrum_id_column_name, ms2q_model_file_name = file_names
     classifiers_file_location = create_test_classifier_csv_file(tmp_path)
 
-    test_library = MS2Library(sqlite_file_loc,
-                              spec2vec_model_file_loc,
-                              ms2ds_model_file_name,
-                              s2v_pickled_embeddings_file,
-                              ms2ds_embeddings_file_name,
-                              ms2q_model_file_name,
-                              classifier_csv_file=classifiers_file_location,
+    test_library = MS2Library(sqlite_file_loc, spec2vec_model_file_loc, ms2ds_model_file_name,
+                              s2v_pickled_embeddings_file, ms2ds_embeddings_file_name, ms2q_model_file_name,
+                              classifier_csv_file_name=classifiers_file_location,
                               spectrum_id_column_name=spectrum_id_column_name)
     test_library.store_potential_true_matches(test_spectra,
                                               os.path.join(tmp_path, "results"),
@@ -199,13 +186,9 @@ def test_store_potential_true_matches_no_matches_found(file_names, test_spectra,
         spectrum_id_column_name, ms2q_model_file_name = file_names
     classifiers_file_location = create_test_classifier_csv_file(tmp_path)
 
-    test_library = MS2Library(sqlite_file_loc,
-                              spec2vec_model_file_loc,
-                              ms2ds_model_file_name,
-                              s2v_pickled_embeddings_file,
-                              ms2ds_embeddings_file_name,
-                              ms2q_model_file_name,
-                              classifier_csv_file=classifiers_file_location,
+    test_library = MS2Library(sqlite_file_loc, spec2vec_model_file_loc, ms2ds_model_file_name,
+                              s2v_pickled_embeddings_file, ms2ds_embeddings_file_name, ms2q_model_file_name,
+                              classifier_csv_file_name=classifiers_file_location,
                               spectrum_id_column_name=spectrum_id_column_name)
     test_library.store_potential_true_matches(test_spectra,
                                               os.path.join(tmp_path, "results"),
@@ -222,12 +205,8 @@ def test_analog_search(file_names, test_spectra):
         ms2ds_model_file_name, ms2ds_embeddings_file_name, \
         spectrum_id_column_name, ms2q_model_file_name = file_names
 
-    test_library = MS2Library(sqlite_file_loc,
-                              spec2vec_model_file_loc,
-                              ms2ds_model_file_name,
-                              s2v_pickled_embeddings_file,
-                              ms2ds_embeddings_file_name,
-                              ms2q_model_file_name,
+    test_library = MS2Library(sqlite_file_loc, spec2vec_model_file_loc, ms2ds_model_file_name,
+                              s2v_pickled_embeddings_file, ms2ds_embeddings_file_name, ms2q_model_file_name,
                               spectrum_id_column_name=spectrum_id_column_name)
 
     cutoff = 20
@@ -246,12 +225,8 @@ def test_calculate_scores_for_metadata(file_names, test_spectra):
         ms2ds_model_file_name, ms2ds_embeddings_file_name, \
         spectrum_id_column_name, ms2q_model_file_name = file_names
 
-    test_library = MS2Library(sqlite_file_loc,
-                              spec2vec_model_file_loc,
-                              ms2ds_model_file_name,
-                              s2v_pickled_embeddings_file,
-                              ms2ds_embeddings_file_name,
-                              ms2q_model_file_name,
+    test_library = MS2Library(sqlite_file_loc, spec2vec_model_file_loc, ms2ds_model_file_name,
+                              s2v_pickled_embeddings_file, ms2ds_embeddings_file_name, ms2q_model_file_name,
                               spectrum_id_column_name=spectrum_id_column_name)
 
     ms2dscores:pd.DataFrame = load_pickled_file(os.path.join(
@@ -276,12 +251,8 @@ def test_get_all_ms2ds_scores(file_names, test_spectra):
         ms2ds_model_file_name, ms2ds_embeddings_file_name, \
         spectrum_id_column_name, ms2q_model_file_name = file_names
 
-    test_library = MS2Library(sqlite_file_loc,
-                              spec2vec_model_file_loc,
-                              ms2ds_model_file_name,
-                              s2v_pickled_embeddings_file,
-                              ms2ds_embeddings_file_name,
-                              ms2q_model_file_name,
+    test_library = MS2Library(sqlite_file_loc, spec2vec_model_file_loc, ms2ds_model_file_name,
+                              s2v_pickled_embeddings_file, ms2ds_embeddings_file_name, ms2q_model_file_name,
                               spectrum_id_column_name=spectrum_id_column_name)
 
     result = test_library._get_all_ms2ds_scores(test_spectra)
@@ -299,12 +270,8 @@ def test_get_s2v_scores(file_names, test_spectra):
         ms2ds_model_file_name, ms2ds_embeddings_file_name, \
         spectrum_id_column_name, ms2q_model_file_name = file_names
 
-    test_library = MS2Library(sqlite_file_loc,
-                              spec2vec_model_file_loc,
-                              ms2ds_model_file_name,
-                              s2v_pickled_embeddings_file,
-                              ms2ds_embeddings_file_name,
-                              ms2q_model_file_name,
+    test_library = MS2Library(sqlite_file_loc, spec2vec_model_file_loc, ms2ds_model_file_name,
+                              s2v_pickled_embeddings_file, ms2ds_embeddings_file_name, ms2q_model_file_name,
                               spectrum_id_column_name=spectrum_id_column_name)
     result = test_library._get_s2v_scores(
         test_spectra[0], ["CCMSLIB00000001572", "CCMSLIB00000001648"])
@@ -317,12 +284,8 @@ def test_get_average_ms2ds_for_inchikey14(file_names):
         ms2ds_model_file_name, ms2ds_embeddings_file_name, \
         spectrum_id_column_name, ms2q_model_file_name = file_names
 
-    test_library = MS2Library(sqlite_file_loc,
-                              spec2vec_model_file_loc,
-                              ms2ds_model_file_name,
-                              s2v_pickled_embeddings_file,
-                              ms2ds_embeddings_file_name,
-                              ms2q_model_file_name,
+    test_library = MS2Library(sqlite_file_loc, spec2vec_model_file_loc, ms2ds_model_file_name,
+                              s2v_pickled_embeddings_file, ms2ds_embeddings_file_name, ms2q_model_file_name,
                               spectrum_id_column_name=spectrum_id_column_name)
     inchickey14s = {"BKUKTJSDOUXYFL", "BTVYFIMKUHNOBZ"}
     ms2ds_scores = pd.Series(
@@ -341,12 +304,8 @@ def test_get_chemical_neighbourhood_scores(file_names):
         ms2ds_model_file_name, ms2ds_embeddings_file_name, \
         spectrum_id_column_name, ms2q_model_file_name = file_names
 
-    test_library = MS2Library(sqlite_file_loc,
-                              spec2vec_model_file_loc,
-                              ms2ds_model_file_name,
-                              s2v_pickled_embeddings_file,
-                              ms2ds_embeddings_file_name,
-                              ms2q_model_file_name,
+    test_library = MS2Library(sqlite_file_loc, spec2vec_model_file_loc, ms2ds_model_file_name,
+                              s2v_pickled_embeddings_file, ms2ds_embeddings_file_name, ms2q_model_file_name,
                               spectrum_id_column_name=spectrum_id_column_name)
     average_inchickey_scores = \
         {'BKUKTJSDOUXYFL': (0.8, 3),
@@ -395,12 +354,8 @@ def test_analog_search_store_in_csv(file_names, test_spectra, tmp_path):
         ms2ds_model_file_name, ms2ds_embeddings_file_name, \
         spectrum_id_column_name, ms2q_model_file_name = file_names
 
-    test_library = MS2Library(sqlite_file_loc,
-                              spec2vec_model_file_loc,
-                              ms2ds_model_file_name,
-                              s2v_pickled_embeddings_file,
-                              ms2ds_embeddings_file_name,
-                              ms2q_model_file_name,
+    test_library = MS2Library(sqlite_file_loc, spec2vec_model_file_loc, ms2ds_model_file_name,
+                              s2v_pickled_embeddings_file, ms2ds_embeddings_file_name, ms2q_model_file_name,
                               spectrum_id_column_name=spectrum_id_column_name)
     results_csv_file = os.path.join(tmp_path, "test_csv_analog_search")
     test_library.analog_search_store_in_csv(test_spectra, results_csv_file)
@@ -411,3 +366,20 @@ def test_analog_search_store_in_csv(file_names, test_spectra, tmp_path):
             '0,905.9927235480093,0.5706255,466.200724,HKSZLNNOFSGOKW,CCMSLIB00000001655,Staurosporine\n',
             '0,926.9927235480093,0.5717702,736.240782,HEWGADDUUGVTPF,CCMSLIB00000001640,Antanapeptin A\n'], \
             "Expected different results to be stored in csv file"
+
+
+def test_create_library_object_from_one_dir():
+    """Test creating a MS2Library object with create_library_object_from_one_dir"""
+    path_to_tests_dir = os.path.join(
+        os.path.split(os.path.dirname(__file__))[0],
+        'tests/test_files/')
+    file_names_dict = {"sqlite": "general_test_files/100_test_spectra.sqlite",
+                       "classifiers": None,
+                       "s2v_model": "general_test_files/100_test_spectra_s2v_model.model",
+                       "ms2ds_model": "general_test_files/ms2ds_siamese_210301_5000_500_400.hdf5",
+                       "ms2query_model": "test_files_ms2library/ms2query_model_all_scores_dropout_regularization.hdf5",
+                       "s2v_embeddings": "general_test_files/100_test_spectra_s2v_embeddings.pickle",
+                       "ms2ds_embeddings": "general_test_files/100_test_spectra_ms2ds_embeddings.pickle"}
+    library = create_library_object_from_one_dir(path_to_tests_dir,
+                                                 file_names_dict)
+    assert isinstance(library, MS2Library)